# Comparing performance - Loop & Single Expression 