Ran twice. It seems that the single expression is significantly faster.  

## Output 1: 
```
goos: linux
goarch: amd64
BenchmarkSingleExpression-12    	1000000000	         0.494 ns/op
BenchmarkLoop-12                	84505893	        13.7 ns/op
PASS
ok  	_/home/quin/tgpl-chapter2/ch2/exercise2-3	1.723s

```


## Output 2: 

```
goos: linux
goarch: amd64
BenchmarkSingleExpression-12    	1000000000	         0.524 ns/op
BenchmarkLoop-12                	84073848	        14.2 ns/op
PASS
ok  	_/home/quin/tgpl-chapter2/ch2/exercise2-3	1.797s
```